[ImageProvider/Loader*] Reorganize in proper class with declarations separated from definitions -> faster compiling, remove warnings, remove stale code in compiled binary

+++ cplusplus/imageprovider/imageproviderhistogram.h
@@ -52,7 +52,7 @@ private:
     QString filepath;
     bool colorversion;
     QImage histimg;
-    int greatestvalue;
+    double greatestvalue;
 
 };
